Larger refactoring of the verification system
