+++ MetaActions.Train/RunReport.cs
@@ -5,14 +5,12 @@
         public string TaskID { get; set; }
         public int TotalMetaActions { get; set; }
         public int TotalValidMetaActions { get; set; }
-        public int TotalUsefulMetaActions { get; set; }
 
-        public RunReport(string taskID, int totalMetaActions, int totalValidMetaActions, int totalUsefulMetaActions)
+        public RunReport(string taskID, int totalMetaActions, int totalValidMetaActions)
         {
             TaskID = taskID;
             TotalMetaActions = totalMetaActions;
             TotalValidMetaActions = totalValidMetaActions;
-            TotalUsefulMetaActions = totalUsefulMetaActions;
         }
     }
 }
