+++ MetaActions.Train/Trainers/MARMATrainer.cs
@@ -14,16 +14,13 @@ using System.Text;
 using System.Threading.Tasks;
 using Tools;
 using MetaActions.Train.MetaActionStrategies;
-using static MetaActions.Learn.Options;
+using static MetaActions.Options;
+using MetaActions.Train.VerificationStrategies;
 
 namespace MetaActions.Train.Trainers
 {
     public class MARMATrainer : BaseCancelable, ITrainer
     {
-        internal string _tempCompiledPath = "compiled";
-        internal string _tempVerificationPath = "verification";
-        internal string _tempReplacementsPath = "replacements";
-
         internal string _outTestProblems = "problems";
         internal string _outCache = "cache";
 
@@ -33,14 +30,12 @@ namespace MetaActions.Train.Trainers
         public TimeSpan TimeLimit { get; } = TimeSpan.FromHours(2);
         public string TempPath { get; }
         public string OutPath { get; }
-        public bool OnlyUsefuls { get; }
         public IMetaActionStrategy MetaActionStrategy { get; }
+        public IVerificationStrategy MetaActionVerificationStrategy { get; }
 
-        internal List<FileInfo> _currentMetaActions = new List<FileInfo>();
-        internal int _totalMetaActions = 0;
         private bool _isDone = false;
 
-        public MARMATrainer(string domainName, FileInfo domain, List<FileInfo> trainingProblems, List<FileInfo> testingProblems, TimeSpan timeLimit, string tempPath, string outPath, bool usefuls, MetaActionStrategy metaActionStrategy) : base(domainName, 1, new CancellationTokenSource())
+        public MARMATrainer(string domainName, FileInfo domain, List<FileInfo> trainingProblems, List<FileInfo> testingProblems, TimeSpan timeLimit, string tempPath, string outPath, MetaActionVerificationStrategy metaActionVerificationStrategy, MetaActionGenerationStrategy metaActionStrategy) : base(domainName, 1, new CancellationTokenSource())
         {
             var _runID = GetDeterministicHashCode(domain.FullName).GetHashCode() + GetDeterministicHashCode(GetType().Name);
             foreach (var trainProblem in trainingProblems)
@@ -53,34 +48,34 @@ namespace MetaActions.Train.Trainers
             TimeLimit = timeLimit;
             TempPath = PathHelper.RootPath(tempPath);
             OutPath = PathHelper.RootPath(outPath);
-            OnlyUsefuls = usefuls;
 
             PathHelper.RecratePath(TempPath);
             PathHelper.RecratePath(OutPath);
 
-            _tempCompiledPath = Path.Combine(TempPath, _tempCompiledPath);
-            _tempVerificationPath = Path.Combine(TempPath, _tempVerificationPath);
-            _tempReplacementsPath = Path.Combine(_tempVerificationPath, _tempReplacementsPath);
-
             _outTestProblems = Path.Combine(OutPath, _outTestProblems);
             _outCache = Path.Combine(OutPath, _outCache);
-            PathHelper.RecratePath(_tempCompiledPath);
-            PathHelper.RecratePath(_tempVerificationPath);
-            PathHelper.RecratePath(_tempReplacementsPath);
 
             PathHelper.RecratePath(_outTestProblems);
             PathHelper.RecratePath(_outCache);
 
             switch (metaActionStrategy)
             {
-                case Learn.Options.MetaActionStrategy.CSMMacros: MetaActionStrategy = new CSMMacros(domainName, RunID, tempPath, CancellationToken); break;
-                case Learn.Options.MetaActionStrategy.PDDLSharpMacros: MetaActionStrategy = new PDDLSharpMacros(domainName, RunID, tempPath, CancellationToken); break;
+                case MetaActionGenerationStrategy.CSMMacros: MetaActionStrategy = new CSMMacros(domainName, RunID, tempPath, CancellationToken); break;
+                case MetaActionGenerationStrategy.PDDLSharpMacros: MetaActionStrategy = new PDDLSharpMacros(domainName, RunID, tempPath, CancellationToken); break;
                 default:
-                    throw new Exception("Unknown meta action strategy!");
+                    throw new Exception("Unknown meta action generation strategy!");
+            }
+
+            switch (metaActionVerificationStrategy)
+            {
+                case Options.MetaActionVerificationStrategy.Strong: MetaActionVerificationStrategy = new StrongVerificationStrategy(domainName, RunID, tempPath, CancellationToken); break;
+                case Options.MetaActionVerificationStrategy.Strong_Useful: MetaActionVerificationStrategy = new StrongUsefulVerificationStrategy(domainName, RunID, tempPath, CancellationToken); break;
+                default:
+                    throw new Exception("Unknown meta action verification strategy!");
             }
 
             CopyTestingProblems(TestingProblems, _outTestProblems);
-            GenerateMetaDomain(Domain, _currentMetaActions, OutPath, TempPath);
+            GenerateMetaDomain(Domain, MetaActionVerificationStrategy.CurrentlyValidMetaActions, OutPath, TempPath);
         }
 
         public Task<RunReport?> RunTask()
@@ -107,76 +102,15 @@ namespace MetaActions.Train.Trainers
             if (CancellationToken.IsCancellationRequested) return null;
 
             Print($"Validating meta actions...", ConsoleColor.Blue);
-            VerifyMetaActions(allMetaActions);
+            var verifiedMetaActions = MetaActionVerificationStrategy.VerifyMetaActions(Domain, allMetaActions, TrainingProblems);
+            Print($"Generating final meta domain...", ConsoleColor.Blue);
+            GenerateMetaDomain(Domain, MetaActionVerificationStrategy.CurrentlyValidMetaActions, OutPath, TempPath);
             if (CancellationToken.IsCancellationRequested) return null;
 
             if (!CancellationToken.IsCancellationRequested)
                 _isDone = true;
 
-            return new RunReport(Name, allMetaActions.Count, _totalMetaActions, _currentMetaActions.Count);
-        }
-
-        internal void VerifyMetaActions(List<FileInfo> allMetaActions)
-        {
-            _totalMetaActions = 0;
-            int metaActionCounter = 1;
-            foreach (var metaAction in allMetaActions)
-            {
-                if (CancellationToken.IsCancellationRequested)
-                    return;
-                PathHelper.RecratePath(_tempReplacementsPath);
-                Print($"\tTesting meta action {metaActionCounter} of {allMetaActions.Count} [{Math.Round(metaActionCounter / (double)allMetaActions.Count * 100, 0)}%]", ConsoleColor.Magenta);
-                int problemCounter = 1;
-                bool allValid = true;
-                foreach (var problem in TrainingProblems)
-                {
-                    if (CancellationToken.IsCancellationRequested)
-                        return;
-                    Print($"\t\tProblem {problemCounter} out of {TrainingProblems.Count} [{Math.Round(problemCounter / (double)TrainingProblems.Count * 100, 0)}%].", ConsoleColor.DarkMagenta);
-                    // Compile Meta Actions
-                    Print($"\t\tCompiling meta action.", ConsoleColor.DarkMagenta);
-                    CompileMetaAction(Domain.FullName, problem.FullName, metaAction.FullName);
-
-                    // Verify Meta Actions
-                    Print($"\t\tVerifying meta action.", ConsoleColor.DarkMagenta);
-                    var isMetaActionValid = VerifyMetaAction();
-
-                    // Stop if invalid
-                    if (!isMetaActionValid)
-                    {
-                        Print($"\tMeta action was invalid in problem '{problem.Name}'.", ConsoleColor.Red);
-                        allValid = false;
-                        break;
-                    }
-                    problemCounter++;
-                }
-                if (allValid)
-                {
-                    _totalMetaActions++;
-                    Print($"\tMeta action was valid in all {TrainingProblems.Count} problems.", ConsoleColor.Green);
-
-                    if (OnlyUsefuls)
-                    {
-                        Print($"\tGenerating initial meta domain...", ConsoleColor.Magenta);
-                        GenerateMetaDomain(Domain, new List<FileInfo>() { metaAction }, OutPath, TempPath);
-
-                        Print("\tChecking for meta action usefulness...", ConsoleColor.Magenta);
-                        if (!IsMetaActionUseful(metaAction, TrainingProblems, TempPath))
-                            continue;
-                        Print("\tMeta Action is Useful!", ConsoleColor.Green);
-                    }
-
-                    _currentMetaActions.Add(metaAction);
-                    Print($"\tExtracting macros from plans...", ConsoleColor.Magenta);
-
-                    ExtractMacrosFromPlans(Domain, _tempReplacementsPath, _outCache);
-                }
-                metaActionCounter++;
-            }
-            Print($"A total of {_currentMetaActions.Count} valid meta actions out of {allMetaActions.Count} was found.", ConsoleColor.Green);
-
-            Print($"Generating final meta domain...", ConsoleColor.Blue);
-            GenerateMetaDomain(Domain, _currentMetaActions, OutPath, TempPath);
+            return new RunReport(Name, allMetaActions.Count, MetaActionVerificationStrategy.CurrentlyValidMetaActions.Count);
         }
 
         public new void Kill()
@@ -186,7 +120,7 @@ namespace MetaActions.Train.Trainers
             if (!_isDone)
             {
                 Print($"Cancelation requested, saving final version of meta domain...", ConsoleColor.Yellow);
-                GenerateMetaDomain(Domain, _currentMetaActions, OutPath, TempPath);
+                GenerateMetaDomain(Domain, MetaActionVerificationStrategy.CurrentlyValidMetaActions, OutPath, TempPath);
             }
         }
 
@@ -221,39 +155,7 @@ namespace MetaActions.Train.Trainers
             }
         }
 
-        internal void CompileMetaAction(string domain, string problem, string metaAction)
-        {
-            ArgsCaller stackelCompiler = ArgsCallerBuilder.GetDotnetRunner("StackelbergCompiler");
-            _activeProcess = stackelCompiler.Process;
-            stackelCompiler.Arguments.Add("--domain", domain);
-            stackelCompiler.Arguments.Add("--problem", problem);
-            stackelCompiler.Arguments.Add("--meta-action", metaAction);
-            stackelCompiler.Arguments.Add("--output", _tempCompiledPath);
-            if (stackelCompiler.Run() != 0 && !CancellationToken.IsCancellationRequested)
-            {
-                Print("Stackelberg Compilation failed!", ConsoleColor.Red);
-                CancellationToken.Cancel();
-            }
-        }
-
-        internal bool VerifyMetaAction()
-        {
-            ArgsCaller stackelVerifier = ArgsCallerBuilder.GetDotnetRunner("StackelbergVerifier");
-            _activeProcess = stackelVerifier.Process;
-            stackelVerifier.Arguments.Add("--domain", Path.Combine(_tempCompiledPath, "simplified_domain.pddl"));
-            stackelVerifier.Arguments.Add("--problem", Path.Combine(_tempCompiledPath, "simplified_problem.pddl"));
-            stackelVerifier.Arguments.Add("--output", _tempVerificationPath);
-            stackelVerifier.Arguments.Add("--iseasy", "");
-            var code = stackelVerifier.Run();
-            if (code != 0 && code != 1 && !CancellationToken.IsCancellationRequested)
-            {
-                Print("Stackelberg Verification failed!", ConsoleColor.Red);
-                CancellationToken.Cancel();
-            }
-            return code == 0;
-        }
-
-        internal void GenerateMetaDomain(FileInfo domainFile, List<FileInfo> metaActionFiles, string outFolder, string tempFolder)
+        internal void GenerateMetaDomain(FileInfo domainFile, List<ValidMetaAction> metaActionFiles, string outFolder, string tempFolder)
         {
             IErrorListener listener = new ErrorListener();
             IParser<INode> parser = new PDDLParser(listener);
@@ -267,66 +169,11 @@ namespace MetaActions.Train.Trainers
             {
                 if (CancellationToken.IsCancellationRequested)
                     return;
-                var metaAction = parser.ParseAs<ActionDecl>(file);
+                var metaAction = parser.ParseAs<ActionDecl>(file.MetaAction);
                 domain.Actions.Add(metaAction);
             }
             generator.Generate(domain, Path.Combine(outFolder, "metaDomain.pddl"));
             generator.Generate(domain, Path.Combine(tempFolder, "metaDomain.pddl"));
-        }
-
-        internal void ExtractMacrosFromPlans(FileInfo domain, string macroPlans, string outFolder)
-        {
-            ArgsCaller macroExtractor = ArgsCallerBuilder.GetDotnetRunner("MacroExtractor");
-            _activeProcess = macroExtractor.Process;
-            macroExtractor.Arguments.Add("--domain", domain.FullName);
-            string macroPlansStr = "";
-            var planFiles = new DirectoryInfo(macroPlans).GetFiles();
-            if (planFiles.Count() == 0 && !CancellationToken.IsCancellationRequested)
-                throw new Exception("Error, there where no plans made from the stackelberg planner");
-            foreach (var plan in planFiles)
-                macroPlansStr += $" {plan.FullName}";
-            macroExtractor.Arguments.Add("--follower-plans", macroPlansStr);
-            macroExtractor.Arguments.Add("--output", outFolder);
-            if (macroExtractor.Run() != 0 && !CancellationToken.IsCancellationRequested)
-            {
-                Print("Macro Extractor failed!", ConsoleColor.Red);
-                CancellationToken.Cancel();
-            }
-        }
-
-        internal bool IsMetaActionUseful(FileInfo metaAction, List<FileInfo> problems, string tempFolder)
-        {
-            var listener = new ErrorListener();
-            var planParser = new FDPlanParser(listener);
-
-            int counter = 1;
-            foreach (var problem in problems)
-            {
-                if (CancellationToken.IsCancellationRequested)
-                    return false;
-                Print($"\t\tUseful check on problem '{problem.Name}' [{counter++}/{problems.Count}]", ConsoleColor.DarkMagenta);
-
-                using (ArgsCaller fdCaller = new ArgsCaller("python3"))
-                {
-                    fdCaller.StdOut += (s, o) => { };
-                    fdCaller.StdErr += (s, o) => { };
-                    fdCaller.Arguments.Add(PathHelper.RootPath("Dependencies/fast-downward/fast-downward.py"), "");
-                    fdCaller.Arguments.Add("--alias", "lama-first");
-                    fdCaller.Arguments.Add("--overall-time-limit", "5m");
-                    fdCaller.Arguments.Add("--plan-file", "plan.plan");
-                    fdCaller.Arguments.Add("metaDomain.pddl", "");
-                    fdCaller.Arguments.Add(problem.FullName, "");
-                    fdCaller.Process.StartInfo.WorkingDirectory = tempFolder;
-                    if (fdCaller.Run() == 0)
-                    {
-                        var plan = planParser.Parse(new FileInfo(Path.Combine(tempFolder, "plan.plan")));
-                        if (plan.Plan.Any(y => y.ActionName == metaAction.Name.Replace(metaAction.Extension, "")))
-                            return true;
-                    }
-                }
-            }
-
-            return false;
         }
     }
 }
