+++ useful-check/plan.plan
@@ -1,0 +1,3 @@
+(down f2 f1)
+($meta_17 f1 p1)
+; cost = 2 (unit cost)
